newsimages25/workflows/SyntaxError404_RETRIEVAL/requirements.txt
torch
numpy
pandas
Pillow
tqdm
ftfy
regex
transformers
